requirements.txt
xarray==2023.6.0
numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.1
cartopy==0.21.1
metpy==1.5.1
Pillow==10.0.0
